+++ lib/init.js
@@ -12,7 +12,7 @@ const loadMongoHistory = utils.loadMongoHistory;
 const loadWSAPI = utils.loadWSAPI;
 const loadQueues = utils.loadQueues;
 const loadServers = utils.loadServers;
-const runnertyio = require('./runnertyio.js');
+const runnertyio = require('./classes/runnertyio.js');
 
 const FilePlan = require('./classes/filePlan.js');
